fix: move about struct
