+++ Cargo.toml
@@ -40,6 +40,7 @@ desktop = [
     "process",
     "dep:freedesktop-desktop-entry",
     "dep:mime",
+    "dep:open",
     "dep:shlex",
     "tokio?/io-util",
     "tokio?/net",
@@ -97,7 +98,7 @@ image = { version = "0.25.1", optional = true }
 lazy_static = "1.4.0"
 libc = { version = "0.2.155", optional = true }
 mime = { version = "0.3.17", optional = true }
-open = "5.3.0"
+open = { version = "5.3.0", optional = true }
 palette = "0.7.3"
 rfd = { version = "0.14.0", optional = true }
 rustix = { version = "0.38.34", features = [
