+++ src/widget/mod.rs
@@ -361,6 +361,3 @@ pub mod tooltip {
 pub mod warning;
 #[doc(inline)]
 pub use warning::*;
-
-pub mod about;
-pub use about::*;
